+++ modules/backend/assets/js/october.sidepaneltab.js
@@ -31,6 +31,10 @@
         $('.fix-button-container', this.$el).append(this.$fixButton)
 
         this.$sideNavItems.click(function(){
+            if ($(this).data('no-side-panel')) {
+                return
+            }
+
             if (Modernizr.touch && $(window).width() < self.options.breakpoint) {
                 if ($(this).data('menu-item') == self.visibleItemId && self.panelVisible) {
                     self.hideSidePanel()
@@ -45,13 +49,16 @@
         })
 
         if (!Modernizr.touch) {
-            self.$sideNav.mouseenter(function(){
-               if ($(window).width() < self.options.breakpoint || !self.panelFixed()) {
-                    self.panelOpenTimeout = setTimeout(function () {
-                        self.displaySidePanel()
-                    }, self.tabOpenDelay)
-               }
-            })
+            // The side panel now opens only when a menu item is hovered and
+            // when the item doesn't have the "data-no-side-panel" attribute.
+            // TODO: remove the comment and the code below if no issues noticed.
+            // self.$sideNav.mouseenter(function(){
+            //     if ($(window).width() < self.options.breakpoint || !self.panelFixed()) {
+            //         self.panelOpenTimeout = setTimeout(function () {
+            //             self.displaySidePanel()
+            //         }, self.tabOpenDelay)
+            //     }
+            // })
 
             self.$sideNav.mouseleave(function(){
                 clearTimeout(self.panelOpenTimeout)
@@ -63,8 +70,14 @@
 
             self.$sideNavItems.mouseenter(function(){
                 if ($(window).width() < self.options.breakpoint || !self.panelFixed()) {
+                    if ($(this).data('no-side-panel')) {
+                        self.hideSidePanel()
+                        return
+                    }
+
                     var _this = this
                     self.tabOpenTimeout = setTimeout(function () {
+                        self.displaySidePanel()
                         self.displayTab(_this)
                     }, self.tabOpenDelay)
                 }
@@ -73,7 +86,6 @@
             self.$sideNavItems.mouseleave(function (){
                 clearTimeout(self.tabOpenTimeout)
             })
-
 
             $(window).resize(function() {
                 self.updatePanelPosition()
@@ -198,7 +210,7 @@
             var data = $this.data('oc.sidePanelTab')
             var options = $.extend({}, SidePanelTab.DEFAULTS, $this.data(), typeof option == 'object' && option)
             if (!data) $this.data('oc.sidePanelTab', (data = new SidePanelTab(this, options)))
-            if (typeof option == 'string') data[option].call($this)
+            if (typeof option == 'string') data[option].call(data)
         })
     }