+++ modules/backend/controllers/Auth.php
@@ -63,8 +63,8 @@ class Auth extends Controller
     public function signin_onSubmit()
     {
         $rules = [
-            'login'    => 'required|min:2|max:32',
-            'password' => 'required|min:4'
+            'login'    => 'required|between:2,255',
+            'password' => 'required|between:4,255'
         ];
 
         $validation = Validator::make(post(), $rules);
@@ -115,7 +115,7 @@ class Auth extends Controller
     public function restore_onSubmit()
     {
         $rules = [
-            'login' => 'required|min:2|max:32'
+            'login' => 'required|between:2,255'
         ];
 
         $validation = Validator::make(post(), $rules);
@@ -176,7 +176,7 @@ class Auth extends Controller
         }
 
         $rules = [
-            'password' => 'required|min:4'
+            'password' => 'required|between:4,255'
         ];
 
         $validation = Validator::make(post(), $rules);