+++ modules/backend/widgets/table/assets/js/build.js
@@ -15,6 +15,7 @@
 =require table.processor.string.js
 =require table.processor.checkbox.js
 =require table.processor.dropdown.js
+=require table.processor.autocomplete.js
 =require table.validator.base.js
 =require table.validator.required.js
 =require table.validator.basenumber.js