+++ modules/system/ServiceProvider.php
@@ -457,7 +457,7 @@ class ServiceProvider extends ModuleServiceProvider
          * - extensions: png,jpg,txt
          */
         Validator::extend('extensions', function($attribute, $value, $parameters) {
-            $extension = $value->getClientOriginalExtension();
+            $extension = strtolower($value->getClientOriginalExtension());
             return in_array($extension, $parameters);
         });