+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* **Build 313** (2015-12-12)
+  - Added a `clickable` option to list columns (see Backend > Lists docs).
+
 * **Build 310** (2015-11-28)
   - The Relation form widget now supports custom SQL with the `select` options, also `nameFrom` supports taking values from model attributes (see Backend > Forms docs).
   - Added Refresh buttons to the back-end logging pages.